Remove Actions REST enterprise subcategory (moved to Actions page) (#25065)
