+++ content/actions/hosting-your-own-runners/autoscaling-with-self-hosted-runners.md
@@ -96,6 +96,6 @@ To  authenticate using a {% data variables.product.prodname_dotcom %} App, it mu
 - For repositories, assign the `administration` permission.
 - For organizations, assign the `organization_self_hosted_runners` permission.
 
-You can register and delete enterprise self-hosted runners using [the API](/rest/reference/enterprise-admin#github-actions). To authenticate to the API, your autoscaling implementation can use an access token.
+You can register and delete enterprise self-hosted runners using [the API](/rest/reference/actions#self-hosted-runners). To authenticate to the API, your autoscaling implementation can use an access token.
 
 Your access token will require the `manage_runners:enterprise` scope.
