+++ lib/redirects/static/client-side-rest-api-redirects.json
@@ -120,5 +120,36 @@
   "/rest/reference/repos#update-a-webhook-configuration-for-a-repository": "/rest/reference/webhooks#update-a-webhook-configuration-for-a-repository",
   "/rest/reference/repos#get-a-delivery-for-a-repository-webhook": "/rest/reference/webhooks#get-a-delivery-for-a-repository-webhook",
   "/rest/reference/repos#list-deliveries-for-a-repository-webhook": "/rest/reference/webhooks#list-deliveries-for-a-repository-webhook",
-  "/rest/reference/repos#redeliver-a-delivery-for-a-repository-webhook": "/rest/reference/webhooks#redeliver-a-delivery-for-a-repository-webhook"
+  "/rest/reference/repos#redeliver-a-delivery-for-a-repository-webhook": "/rest/reference/webhooks#redeliver-a-delivery-for-a-repository-webhook",
+  "/rest/reference/enterprise-admin#add-custom-labels-to-a-self-hosted-runner-for-an-enterprise": "/rest/reference/actions#add-custom-labels-to-a-self-hosted-runner-for-an-enterprise",
+  "/rest/reference/enterprise-admin#add-organization-access-to-a-self-hosted-runner-group-in-an-enterprise": "/rest/reference/actions#add-organization-access-to-a-self-hosted-runner-group-in-an-enterprise",
+  "/rest/reference/enterprise-admin#create-a-registration-token-for-an-enterprise": "/rest/reference/actions#create-a-registration-token-for-an-enterprise",
+  "/rest/reference/enterprise-admin#create-a-remove-token-for-an-enterprise": "/rest/reference/actions#create-a-remove-token-for-an-enterprise",
+  "/rest/reference/enterprise-admin#create-self-hosted-runner-group-for-an-enterprise": "/rest/reference/actions#create-self-hosted-runner-group-for-an-enterprise",
+  "/rest/reference/enterprise-admin#delete-self-hosted-runner-from-an-enterprise": "/rest/reference/actions#delete-self-hosted-runner-from-an-enterprise",
+  "/rest/reference/enterprise-admin#delete-a-self-hosted-runner-group-from-an-enterprise": "/rest/reference/actions#delete-a-self-hosted-runner-group-from-an-enterprise",
+  "/rest/reference/enterprise-admin#disable-a-selected-organization-for-github-actions-in-an-enterprise": "/rest/reference/actions#disable-a-selected-organization-for-github-actions-in-an-enterprise",
+  "/rest/reference/enterprise-admin#enable-a-selected-organization-for-github-actions-in-an-enterprise": "/rest/reference/actions#enable-a-selected-organization-for-github-actions-in-an-enterprise",
+  "/rest/reference/enterprise-admin#get-allowed-actions-for-an-enterprise": "/rest/reference/actions#get-allowed-actions-for-an-enterprise",
+  "/rest/reference/enterprise-admin#get-github-actions-permissions-for-an-enterprise": "/rest/reference/actions#get-github-actions-permissions-for-an-enterprise",
+  "/rest/reference/enterprise-admin#get-a-self-hosted-runner-for-an-enterprise": "/rest/reference/actions#get-a-self-hosted-runner-for-an-enterprise",
+  "/rest/reference/enterprise-admin#get-a-self-hosted-runner-group-for-an-enterprise": "/rest/reference/actions#get-a-self-hosted-runner-group-for-an-enterprise",
+  "/rest/reference/enterprise-admin#list-labels-for-a-self-hosted-runner-for-an-enterprise": "/rest/reference/actions#list-labels-for-a-self-hosted-runner-for-an-enterprise",
+  "/rest/reference/enterprise-admin#list-organization-access-to-a-self-hosted-runner-group-in-a-enterprise": "/rest/reference/actions#list-organization-access-to-a-self-hosted-runner-group-in-a-enterprise",
+  "/rest/reference/enterprise-admin#list-runner-applications-for-an-enterprise": "/rest/reference/actions#list-runner-applications-for-an-enterprise",
+  "/rest/reference/enterprise-admin#list-selected-organizations-enabled-for-github-actions-in-an-enterprise": "/rest/reference/actions#list-selected-organizations-enabled-for-github-actions-in-an-enterprise",
+  "/rest/reference/enterprise-admin#list-self-hosted-runner-groups-for-an-enterprise": "/rest/reference/actions#list-self-hosted-runner-groups-for-an-enterprise",
+  "/rest/reference/enterprise-admin#list-self-hosted-runners-for-an-enterprise": "/rest/reference/actions#list-self-hosted-runners-for-an-enterprise",
+  "/rest/reference/enterprise-admin#list-self-hosted-runners-in-a-group-for-an-enterprise": "/rest/reference/actions#list-self-hosted-runners-in-a-group-for-an-enterprise",
+  "/rest/reference/enterprise-admin#remove-all-custom-labels-from-a-self-hosted-runner-for-an-enterprise": "/rest/reference/actions#remove-all-custom-labels-from-a-self-hosted-runner-for-an-enterprise",
+  "/rest/reference/enterprise-admin#remove-a-custom-label-from-a-self-hosted-runner-for-an-enterprise": "/rest/reference/actions#remove-a-custom-label-from-a-self-hosted-runner-for-an-enterprise",
+  "/rest/reference/enterprise-admin#remove-organization-access-to-a-self-hosted-runner-group-in-an-enterprise": "/rest/reference/actions#remove-organization-access-to-a-self-hosted-runner-group-in-an-enterprise",
+  "/rest/reference/enterprise-admin#remove-a-self-hosted-runner-from-a-group-for-an-enterprise": "/rest/reference/actions#remove-a-self-hosted-runner-from-a-group-for-an-enterprise",
+  "/rest/reference/enterprise-admin#set-allowed-actions-for-an-enterprise": "/rest/reference/actions#set-allowed-actions-for-an-enterprise",
+  "/rest/reference/enterprise-admin#set-custom-labels-for-a-self-hosted-runner-for-an-enterprise": "/rest/reference/actions#set-custom-labels-for-a-self-hosted-runner-for-an-enterprise",
+  "/rest/reference/enterprise-admin#set-github-actions-permissions-for-an-enterprise": "/rest/reference/actions#set-github-actions-permissions-for-an-enterprise",
+  "/rest/reference/enterprise-admin#set-organization-access-to-a-self-hosted-runner-group-in-an-enterprise": "/rest/reference/actions#set-organization-access-to-a-self-hosted-runner-group-in-an-enterprise",
+  "/rest/reference/enterprise-admin#set-selected-organizations-enabled-for-github-actions-in-an-enterprise": "/rest/reference/actions#set-selected-organizations-enabled-for-github-actions-in-an-enterprise",
+  "/rest/reference/enterprise-admin#set-self-hosted-runners-in-a-group-for-an-enterprise": "/rest/reference/actions#set-self-hosted-runners-in-a-group-for-an-enterprise",
+  "/rest/reference/enterprise-admin#update-a-self-hosted-runner-group-for-an-enterprise": "/rest/reference/actions#update-a-self-hosted-runner-group-for-an-enterprise"
 }
